fix narrow

+++ game/shared/tf/tf_shareddefs.cpp
@@ -147,7 +147,7 @@ const char *g_aWeaponNames[] =
 	"TF_WEAPON_COUNT",	// end marker, do not add below here
 };
 
-int g_aWeaponDamageTypes[] =
+unsigned int g_aWeaponDamageTypes[] =
 {
 	DMG_GENERIC,	// TF_WEAPON_NONE
 	DMG_CLUB,		// TF_WEAPON_BAT,
@@ -206,7 +206,7 @@ int g_aWeaponDamageTypes[] =
 
 	// This is a special entry that must match with TF_WEAPON_COUNT
 	// to protect against updating the weapon list without updating this list
-	(unsigned)TF_DMG_SENTINEL_VALUE
+	TF_DMG_SENTINEL_VALUE
 };
 
 const char *g_szProjectileNames[] =
